+++ CHANGELOG.md
@@ -47,6 +47,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Error loading Etoys activity #1717
 - Cheating in Abecedarium Activity #1720
 - X and Y axis labels are not Interchanged when switching between Vertical and Horizontal Bar Charts #1727
+- BlockRain Activity Game Over Logic Fails When Side Columns Are Full #1739
  
 ## [1.8.0] - 2024-04-10
 ### Added